scheduler/requirements.txt
requests>=2.32.0
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
schedule>=1.2.2
tzdata>=2024.1
supabase>=2.9.0
cryptography>=44.0.0
